requirements.txt
Django
onnxruntime
numpy
Pillow
requests
python-dotenv
whitenoise
